Reorganise code for test
Repackaged test classes for Person and Tag.
