+++ src/test/java/seedu/address/logic/parser/tag/CreateTagCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.tag;
 
 import static seedu.address.logic.commands.CommandTestUtil.INVALID_TAG_DESC;
 import static seedu.address.logic.commands.CommandTestUtil.PREAMBLE_WHITESPACE;
@@ -12,6 +12,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.tag.CreateTagCommand;
+import seedu.address.logic.parser.CreateTagCommandParser;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.TagName;
