+++ src/test/java/seedu/address/logic/parser/tag/DeleteTagCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.tag;
 
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
@@ -8,6 +8,7 @@ import static seedu.address.testutil.TypicalTags.FLORIST;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.tag.DeleteTagCommand;
+import seedu.address.logic.parser.DeleteTagCommandParser;
 
 public class DeleteTagCommandParserTest {
