+++ src/test/java/seedu/address/logic/parser/tag/UntagCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.tag;
 
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
@@ -11,6 +11,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.tag.UntagCommand;
+import seedu.address.logic.parser.UntagCommandParser;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.TagName;
